Refactors authentication into separate file.
