+++ example/example.dart
@@ -106,71 +106,4 @@ void main() async {
   print('Client Id: ${credentials.clientId}');
   print('Access Token: ${credentials.accessToken}');
   print('Credentials Expired: ${credentials.isExpired}');
-
-  print('Would you like to access user specific data? [y/n]');
-  var yes = stdin.readLineSync();
-  if ('y' != yes.trim().toLowerCase()) {
-    exit(0);
-    return;
-  }
-  spotify = await _getUserAuthenticatedSpotifyApi(credentials);
-  if (spotify == null) {
-    exit(0);
-    return;
-  }
-  await _currentlyPlaying(spotify);
-  await _devices(spotify);
-
-  exit(0);
-}
-
-Future<SpotifyApi> _getUserAuthenticatedSpotifyApi(
-    SpotifyApiCredentials credentials) async {
-  print(
-      'Please paste your redirect url (from your spotify application\'s redirect url):');
-  var redirect = stdin.readLineSync();
-
-  var grant = SpotifyApi.authorizationCodeGrant(credentials);
-  var authUri = grant.getAuthorizationUrl(Uri.parse(redirect),
-      scopes: ['user-read-playback-state']);
-
-  print(
-      'Please paste this url \n\n$authUri\n\nto your browser and enter the redirected url:');
-  var redirectUrl;
-  var userInput = stdin.readLineSync();
-  if (userInput == null || (redirectUrl = Uri.tryParse(userInput)) == null) {
-    print('Invalid redirect url');
-    return null;
-  }
-
-  var client =
-      await grant.handleAuthorizationResponse(redirectUrl.queryParameters);
-  return SpotifyApi.fromClient(client);
-}
-
-void _currentlyPlaying(SpotifyApi spotify) async =>
-    await spotify.users.currentlyPlaying().then((a) {
-      if (a == null) {
-        print('Nothing currently playing.');
-        return;
-      }
-      print('Currently playing: ${a.item.name}');
-    }).catchError(_prettyPrintError);
-
-void _devices(SpotifyApi spotify) async =>
-    await spotify.users.devices().then((devices) {
-      if (devices == null) {
-        print('No devices currently playing.');
-        return;
-      }
-      print('Listing ${devices.length} available devices:');
-      print(devices.map((device) => device.name).join(', '));
-    }).catchError(_prettyPrintError);
-
-void _prettyPrintError(Exception error) {
-  if (error is SpotifyException) {
-    print('${error.status} : ${error.message}');
-  } else {
-    print(error);
-  }
 }
